requirements.txt
Django>=5.2.4
gunicorn>=22.0.0
whitenoise>=6.7.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
stripe>=11.0.0
